requirements.txt
amqp==2.3.2
anyjson==0.3.3
appdirs==1.4.3
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.1.1
cffi==1.11.5
colorama==0.3.9
configparser==3.5.0
coverage==3.7.1
cryptography==2.2.2
cycler==0.10.0
Cython==0.25.2
Django==1.8.15
django-celery==3.1.17
django-datatables-view==1.13.0
django-filter==0.9.2
django-redis==4.8.0
django-rest-swagger==0.3.4
djangorestframework==3.1.0
docxtpl==0.5.0
dwebsocket==0.4.2
ecdsa==0.13
elasticsearch==2.2.0
enum34==1.1.6
Fabric==1.10.1
fake-factory==0.5.8
Faker==0.8.16
funcsigs==1.0.2
futures==3.2.0
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.8.1
idna==2.7
ipaddress==1.0.22
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
JPype1==0.6.3
kafka==1.3.5
kafka-python==1.3.5
kazoo==2.4.0
kiwisolver==1.0.1
kombu==4.2.1
lazy-object-proxy==1.3.1
lxml==4.2.1
Markdown==2.6
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mock==1.3.0
ndg-httpsclient==0.4.2
netaddr==0.7.18
numpy==1.14.5
oath==1.4.1
packaging==17.1
paramiko==1.15.3
pbr==4.0.4
pdfkit==0.6.1
pep8==1.7.1
Pillow==3.0.0
psutil==5.2.0
py-zabbix==0.6.1
pyasn1==0.4.3
pycodestyle==2.4.0
pycparser==2.18
pycrypto==2.6.1
pykafka==2.7.0
pylint==1.8.2
pymongo==3.7.0
PyMySQL==0.7.11
PyNaCl==1.2.1
pyOpenSSL==16.2.0
pyparsing==2.2.0
pypinyin==0.30.0
python-dateutil==2.4.2
python-docx==0.8.6
pytz==2018.4
pywebhdfs==0.4.1
PyYAML==3.11
pyzmq==13.1.0
qrcode==5.3
redis==2.10.5
reportlab==3.4.0
requests==2.6.0
singledispatch==3.4.0.3
six==1.11.0
tabulate==0.8.2
text-unidecode==1.2
typing==3.6.4
urllib3==1.23
vine==1.1.4
wrapt==1.10.11
xlwt==0.7.5
xmltodict==0.10.1
hdfs==2.1.0